refactor: create styles for improved organization
