+++ mobile/components/PrimaryButton/styles.ts
@@ -1,0 +1,26 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  buttonPrimary: {
+    backgroundColor: '#A4C457',
+    paddingHorizontal: 16,
+    borderRadius: 10,
+    marginBottom: 16,
+    height: 56,
+    width: '100%',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+
+  buttonHover: {
+    backgroundColor: '#8FA743',
+  },
+
+  buttonText: {
+    color: '#fff',
+    fontSize: 16,
+    fontWeight: 'bold',
+    textAlign: 'center',
+    lineHeight: 56,
+  },
+});
